完成订单提交及支付webSocket实现支付回显，ngrok实现内网穿透

+++ beans/src/main/java/com/gonghr/fmmall/entity/Orders.java
@@ -20,7 +20,7 @@ public class Orders {
     private String receiverMobile;
     private String receiverAddress;
     private BigDecimal totalAmount;
-    private Integer actualAmount;
+    private BigDecimal actualAmount;
     private Integer payType;
     private String orderRemark;
     private String status;
